+++ spec/integration/resource/api300/synergy/volume_attachment/update_spec.rb
@@ -1,11 +1,11 @@
 require 'spec_helper'
 
-RSpec.describe OneviewSDK::API300::Thunderbird::VolumeAttachment, integration: true, type: UPDATE do
+RSpec.describe OneviewSDK::API300::Synergy::VolumeAttachment, integration: true, type: UPDATE do
   include_context 'integration api300 context'
 
   describe '#update' do
     it 'should throw unavailable exception' do
-      item = OneviewSDK::API300::Thunderbird::VolumeAttachment.new($client_300)
+      item = OneviewSDK::API300::Synergy::VolumeAttachment.new($client_300_synergy)
       expect { item.update }.to raise_error(OneviewSDK::MethodUnavailable)
     end
   end